move some date functionality into lib
