+++ lib/Basics/datetime.cpp
@@ -21,16 +21,503 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 #include "Basics/datetime.h"
-#include <date/date.h>
 #include "Basics/NumberUtils.h"
 #include "Logger/Logger.h"
 
 #include <boost/algorithm/string.hpp>
+#include <date/date.h>
+#include <date/iso_week.h>
 
+#include <chrono>
 #include <regex>
 #include <vector>
 
 namespace {
+using namespace date;
+using namespace std::chrono;
+
+std::string tail(std::string const& source, size_t const length) {
+  if (length >= source.size()) {
+    return source;
+  }
+  return source.substr(source.size() - length);
+}  // tail
+
+typedef void (*format_func_t)(std::string& wrk, arangodb::tp_sys_clock_ms const&);
+std::unordered_map<std::string, format_func_t> dateMap;
+auto const unixEpoch = date::sys_seconds{std::chrono::seconds{0}};
+
+std::vector<std::string> const monthNames = {"January", "February", "March",
+                                             "April",   "May",      "June",
+                                             "July",    "August",   "September",
+                                             "October", "November", "December"};
+
+std::vector<std::string> const monthNamesShort = {"Jan", "Feb", "Mar", "Apr",
+                                                  "May", "Jun", "Jul", "Aug",
+                                                  "Sep", "Oct", "Nov", "Dec"};
+
+std::vector<std::string> const weekDayNames = {"Sunday",   "Monday",
+                                               "Tuesday",  "Wednesday",
+                                               "Thursday", "Friday",
+                                               "Saturday"};
+
+std::vector<std::string> const weekDayNamesShort = {"Sun", "Mon", "Tue", "Wed",
+                                                    "Thu", "Fri", "Sat"};
+
+std::
+    vector<std::pair<std::string, format_func_t>> const sortedDateMap = {{"%&",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                          }},  // Allow for literal "m" after "%m" ("%mm" ->
+                                                                               // %m%&m)
+                                                                         // zero-pad 4 digit years to length of 6 and add "+"
+                                                                         // prefix, keep negative as-is
+                                                                         {"%yyy"
+                                                                          "yyy",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day(
+                                                                                floor<date::days>(tp));
+                                                                            auto yearnum = static_cast<int>(
+                                                                                ymd.year());
+                                                                            if (yearnum < 0) {
+                                                                              if (yearnum > -10) {
+                                                                                wrk.append(
+                                                                                    "-00000");
+                                                                              } else if (yearnum > -100) {
+                                                                                wrk.append(
+                                                                                    "-0000");
+                                                                              } else if (yearnum > -1000) {
+                                                                                wrk.append(
+                                                                                    "-000");
+                                                                              } else if (yearnum > -10000) {
+                                                                                wrk.append(
+                                                                                    "-00");
+                                                                              } else if (yearnum > -100000) {
+                                                                                wrk.append(
+                                                                                    "-0");
+                                                                              } else {
+                                                                                wrk.append(
+                                                                                    "-");
+                                                                              }
+                                                                              wrk.append(std::to_string(
+                                                                                  abs(yearnum)));
+                                                                              return;
+                                                                            }
+
+                                                                            TRI_ASSERT(yearnum >= 0);
+
+                                                                            if (yearnum > 99999) {
+                                                                              // intentionally nothing
+                                                                            } else if (yearnum > 9999) {
+                                                                              wrk.append(
+                                                                                  "+0");
+                                                                            } else if (yearnum > 999) {
+                                                                              wrk.append(
+                                                                                  "+00");
+                                                                            } else if (yearnum > 99) {
+                                                                              wrk.append(
+                                                                                  "+000");
+                                                                            } else if (yearnum > 9) {
+                                                                              wrk.append(
+                                                                                  "+0000");
+                                                                            } else {
+                                                                              wrk.append(
+                                                                                  "+00000");
+                                                                            }
+                                                                            wrk.append(std::to_string(yearnum));
+                                                                          }},
+                                                                         {"%mmm"
+                                                                          "m",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day(
+                                                                                floor<date::days>(tp));
+                                                                            wrk.append(
+                                                                                ::monthNames[static_cast<unsigned>(ymd.month()) - 1]);
+                                                                          }},
+                                                                         {"%yyy"
+                                                                          "y",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day(
+                                                                                floor<date::days>(tp));
+                                                                            auto yearnum = static_cast<int>(
+                                                                                ymd.year());
+                                                                            if (yearnum < 0) {
+                                                                              if (yearnum > -10) {
+                                                                                wrk.append(
+                                                                                    "-000");
+                                                                              } else if (yearnum > -100) {
+                                                                                wrk.append(
+                                                                                    "-00");
+                                                                              } else if (yearnum > -1000) {
+                                                                                wrk.append(
+                                                                                    "-0");
+                                                                              } else {
+                                                                                wrk.append(
+                                                                                    "-");
+                                                                              }
+                                                                              wrk.append(std::to_string(
+                                                                                  abs(yearnum)));
+                                                                            } else {
+                                                                              TRI_ASSERT(yearnum >= 0);
+                                                                              if (yearnum < 9) {
+                                                                                wrk.append(
+                                                                                    "000");
+                                                                                wrk.append(std::to_string(yearnum));
+                                                                              } else if (yearnum < 99) {
+                                                                                wrk.append(
+                                                                                    "00");
+                                                                                wrk.append(std::to_string(yearnum));
+                                                                              } else if (yearnum < 999) {
+                                                                                wrk.append(
+                                                                                    "0");
+                                                                                wrk.append(std::to_string(yearnum));
+                                                                              } else {
+                                                                                std::string yearstr(
+                                                                                    std::to_string(yearnum));
+                                                                                wrk.append(tail(yearstr, 4));
+                                                                              }
+                                                                            }
+                                                                          }},
+
+                                                                         {"%www"
+                                                                          "w",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            weekday wd{floor<date::days>(tp)};
+                                                                            wrk.append(
+                                                                                ::weekDayNames[static_cast<unsigned>(wd)]);
+                                                                          }},
+
+                                                                         {"%mm"
+                                                                          "m",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day(
+                                                                                floor<date::days>(tp));
+                                                                            wrk.append(
+                                                                                ::monthNamesShort[static_cast<unsigned>(ymd.month()) - 1]);
+                                                                          }},
+                                                                         {"%ww"
+                                                                          "w",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            weekday wd{floor<date::days>(tp)};
+                                                                            wrk.append(
+                                                                                weekDayNamesShort[static_cast<unsigned>(wd)]);
+                                                                          }},
+                                                                         {"%ff"
+                                                                          "f",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto day_time = make_time(
+                                                                                tp - floor<date::days>(tp));
+                                                                            uint64_t millis =
+                                                                                day_time
+                                                                                    .subseconds()
+                                                                                    .count();
+                                                                            if (millis < 10) {
+                                                                              wrk.append(
+                                                                                  "00");
+                                                                            } else if (millis < 100) {
+                                                                              wrk.append(
+                                                                                  "0");
+                                                                            }
+                                                                            wrk.append(std::to_string(millis));
+                                                                          }},
+                                                                         {"%xx"
+                                                                          "x",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day(
+                                                                                floor<date::days>(tp));
+                                                                            auto yyyy =
+                                                                                year{ymd.year()};
+                                                                            // we construct the date with the first day in the
+                                                                            // year:
+                                                                            auto firstDayInYear =
+                                                                                yyyy / jan /
+                                                                                day{0};
+                                                                            uint64_t daysSinceFirst =
+                                                                                duration_cast<date::days>(
+                                                                                    tp - sys_days(firstDayInYear))
+                                                                                    .count();
+                                                                            if (daysSinceFirst < 10) {
+                                                                              wrk.append(
+                                                                                  "00");
+                                                                            } else if (daysSinceFirst < 100) {
+                                                                              wrk.append(
+                                                                                  "0");
+                                                                            }
+                                                                            wrk.append(std::to_string(daysSinceFirst));
+                                                                          }},
+
+                                                                         // there"s no really sensible way to handle negative
+                                                                         // years, but better not drop the sign
+                                                                         {"%yy",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day(
+                                                                                floor<date::days>(tp));
+                                                                            auto yearnum = static_cast<int>(
+                                                                                ymd.year());
+                                                                            if (yearnum < 10 &&
+                                                                                yearnum > -10) {
+                                                                              wrk.append(
+                                                                                  "0");
+                                                                              wrk.append(std::to_string(
+                                                                                  abs(yearnum)));
+                                                                            } else {
+                                                                              std::string yearstr(std::to_string(
+                                                                                  abs(yearnum)));
+                                                                              wrk.append(tail(yearstr, 2));
+                                                                            }
+                                                                          }},
+                                                                         {"%mm",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day(
+                                                                                floor<date::days>(tp));
+                                                                            auto month = static_cast<unsigned>(
+                                                                                ymd.month());
+                                                                            if (month < 10) {
+                                                                              wrk.append(
+                                                                                  "0");
+                                                                            }
+                                                                            wrk.append(std::to_string(month));
+                                                                          }},
+                                                                         {"%dd",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day(
+                                                                                floor<date::days>(tp));
+                                                                            auto day = static_cast<unsigned>(
+                                                                                ymd.day());
+                                                                            if (day < 10) {
+                                                                              wrk.append(
+                                                                                  "0");
+                                                                            }
+                                                                            wrk.append(std::to_string(day));
+                                                                          }},
+                                                                         {"%hh",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto day_time = make_time(
+                                                                                tp - floor<date::days>(tp));
+                                                                            uint64_t hours =
+                                                                                day_time
+                                                                                    .hours()
+                                                                                    .count();
+                                                                            if (hours < 10) {
+                                                                              wrk.append(
+                                                                                  "0");
+                                                                            }
+                                                                            wrk.append(std::to_string(hours));
+                                                                          }},
+                                                                         {"%ii",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto day_time = make_time(
+                                                                                tp - floor<date::days>(tp));
+                                                                            uint64_t minutes =
+                                                                                day_time
+                                                                                    .minutes()
+                                                                                    .count();
+                                                                            if (minutes < 10) {
+                                                                              wrk.append(
+                                                                                  "0");
+                                                                            }
+                                                                            wrk.append(std::to_string(minutes));
+                                                                          }},
+                                                                         {"%ss",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto day_time = make_time(
+                                                                                tp - floor<date::days>(tp));
+                                                                            uint64_t seconds =
+                                                                                day_time
+                                                                                    .seconds()
+                                                                                    .count();
+                                                                            if (seconds < 10) {
+                                                                              wrk.append(
+                                                                                  "0");
+                                                                            }
+                                                                            wrk.append(std::to_string(seconds));
+                                                                          }},
+                                                                         {"%kk",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            iso_week::year_weeknum_weekday yww{
+                                                                                floor<date::days>(tp)};
+                                                                            uint64_t isoWeek =
+                                                                                static_cast<unsigned>(
+                                                                                    yww.weeknum());
+                                                                            if (isoWeek < 10) {
+                                                                              wrk.append(
+                                                                                  "0");
+                                                                            }
+                                                                            wrk.append(std::to_string(isoWeek));
+                                                                          }},
+
+                                                                         {"%t",
+                                                                          [](std::string& wrk, arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto diffDuration =
+                                                                                tp - unixEpoch;
+                                                                            auto diff =
+                                                                                duration_cast<duration<double, std::milli>>(
+                                                                                    diffDuration)
+                                                                                    .count();
+                                                                            wrk.append(std::to_string(static_cast<int64_t>(
+                                                                                std::round(diff))));
+                                                                          }},
+                                                                         {"%z",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            std::string formatted = format(
+                                                                                "%FT%TZ",
+                                                                                floor<milliseconds>(tp));
+                                                                            wrk.append(formatted);
+                                                                          }},
+                                                                         {"%w",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            weekday wd{floor<date::days>(tp)};
+                                                                            wrk.append(std::to_string(
+                                                                                static_cast<unsigned>(wd)));
+                                                                          }},
+                                                                         {"%y",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day(
+                                                                                floor<date::days>(tp));
+                                                                            wrk.append(std::to_string(static_cast<int>(
+                                                                                ymd.year())));
+                                                                          }},
+                                                                         {"%m",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day(
+                                                                                floor<date::days>(tp));
+                                                                            wrk.append(std::to_string(static_cast<unsigned>(
+                                                                                ymd.month())));
+                                                                          }},
+                                                                         {"%d",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day(
+                                                                                floor<date::days>(tp));
+                                                                            wrk.append(std::to_string(static_cast<unsigned>(
+                                                                                ymd.day())));
+                                                                          }},
+                                                                         {"%h",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto day_time = make_time(
+                                                                                tp - floor<date::days>(tp));
+                                                                            uint64_t hours =
+                                                                                day_time
+                                                                                    .hours()
+                                                                                    .count();
+                                                                            wrk.append(std::to_string(hours));
+                                                                          }},
+                                                                         {"%i",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto day_time = make_time(
+                                                                                tp - floor<date::days>(tp));
+                                                                            uint64_t minutes =
+                                                                                day_time
+                                                                                    .minutes()
+                                                                                    .count();
+                                                                            wrk.append(std::to_string(minutes));
+                                                                          }},
+                                                                         {"%s",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto day_time = make_time(
+                                                                                tp - floor<date::days>(tp));
+                                                                            uint64_t seconds =
+                                                                                day_time
+                                                                                    .seconds()
+                                                                                    .count();
+                                                                            wrk.append(std::to_string(seconds));
+                                                                          }},
+                                                                         {"%f",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto day_time = make_time(
+                                                                                tp - floor<date::days>(tp));
+                                                                            uint64_t millis =
+                                                                                day_time
+                                                                                    .subseconds()
+                                                                                    .count();
+                                                                            wrk.append(std::to_string(millis));
+                                                                          }},
+                                                                         {"%x",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day(
+                                                                                floor<date::days>(tp));
+                                                                            auto yyyy =
+                                                                                year{ymd.year()};
+                                                                            // We construct the date with the first day in the
+                                                                            // year:
+                                                                            auto firstDayInYear =
+                                                                                yyyy / jan /
+                                                                                day{0};
+                                                                            uint64_t daysSinceFirst =
+                                                                                duration_cast<date::days>(
+                                                                                    tp - sys_days(firstDayInYear))
+                                                                                    .count();
+                                                                            wrk.append(std::to_string(daysSinceFirst));
+                                                                          }},
+                                                                         {"%k",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            iso_week::year_weeknum_weekday yww{
+                                                                                floor<date::days>(tp)};
+                                                                            uint64_t isoWeek =
+                                                                                static_cast<unsigned>(
+                                                                                    yww.weeknum());
+                                                                            wrk.append(std::to_string(isoWeek));
+                                                                          }},
+                                                                         {"%l",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            year_month_day ymd{
+                                                                                floor<date::days>(tp)};
+                                                                            if (ymd.year()
+                                                                                    .is_leap()) {
+                                                                              wrk.append(
+                                                                                  "1");
+                                                                            } else {
+                                                                              wrk.append(
+                                                                                  "0");
+                                                                            }
+                                                                          }},
+                                                                         {"%q",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            year_month_day ymd{
+                                                                                floor<date::days>(tp)};
+                                                                            month m = ymd.month();
+                                                                            uint64_t part = static_cast<uint64_t>(
+                                                                                ceil(unsigned(m) / 3.0f));
+                                                                            TRI_ASSERT(part <= 4);
+                                                                            wrk.append(std::to_string(part));
+                                                                          }},
+                                                                         {"%a",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            auto ymd = year_month_day{
+                                                                                floor<date::days>(tp)};
+                                                                            auto lastMonthDay =
+                                                                                ymd.year() /
+                                                                                ymd.month() / last;
+                                                                            wrk.append(std::to_string(static_cast<unsigned>(
+                                                                                lastMonthDay
+                                                                                    .day())));
+                                                                          }},
+                                                                         {"%%",
+                                                                          [](std::string& wrk,
+                                                                             arangodb::tp_sys_clock_ms const& tp) {
+                                                                            wrk.append(
+                                                                                "%");
+                                                                          }},
+                                                                         {"%", [](std::string& wrk,
+                                                                                  arangodb::tp_sys_clock_ms const& tp) {
+                                                                          }}};
+
+// will be populated by DateRegexInitializer
+std::regex dateFormatRegex;
+
 std::regex const iso8601Regex(
     "(\\+|\\-)?\\d+(\\-\\d{1,2}(\\-\\d{1,2})?)?"  // YY[YY]-MM-DD
     "("
@@ -92,9 +579,64 @@ std::regex const durationRegex(
     "P((\\d+)Y)?((\\d+)M)?((\\d+)W)?((\\d+)D)?(T((\\d+)H)?((\\d+)M)?((\\d+)(\\."
     "(\\d{1,3}))?S)?)?");
 
+struct DateRegexInitializer {
+  DateRegexInitializer() {
+    std::string myregex;
+
+    dateMap.reserve(sortedDateMap.size());
+    std::for_each(sortedDateMap.begin(), sortedDateMap.end(),
+                  [&myregex](std::pair<std::string const&, format_func_t> const& p) {
+                    (myregex.length() > 0) ? myregex += "|" + p.first : myregex = p.first;
+                    dateMap.insert(std::make_pair(p.first, p.second));
+                  });
+    dateFormatRegex = std::regex(myregex);
+  }
+};
+
+std::string executeDateFormatRegex(std::string const& search, 
+                                   arangodb::tp_sys_clock_ms const& tp) {
+  std::string s;
+
+  auto first = search.begin();
+  auto last = search.end();
+  typename std::smatch::difference_type positionOfLastMatch = 0;
+  auto endOfLastMatch = first;
+
+  auto callback = [&tp, &endOfLastMatch, &positionOfLastMatch, &s](std::smatch const& match) {
+    auto positionOfThisMatch = match.position(0);
+    auto diff = positionOfThisMatch - positionOfLastMatch;
+
+    auto startOfThisMatch = endOfLastMatch;
+    std::advance(startOfThisMatch, diff);
+
+    s.append(endOfLastMatch, startOfThisMatch);
+    auto got = dateMap.find(match.str(0));
+    if (got != dateMap.end()) {
+      got->second(s, tp);
+    }
+    auto lengthOfMatch = match.length(0);
+
+    positionOfLastMatch = positionOfThisMatch + lengthOfMatch;
+
+    endOfLastMatch = startOfThisMatch;
+    std::advance(endOfLastMatch, lengthOfMatch);
+  };
+
+  std::regex_iterator<std::string::const_iterator> end;
+  std::regex_iterator<std::string::const_iterator> begin(first, last, dateFormatRegex);
+  std::for_each(begin, end, callback);
+
+  s.append(endOfLastMatch, last);
+
+  return s;
+}
+
+// populates dateFormatRegex
+static DateRegexInitializer const initializer;
+
 }  // namespace
 
-bool arangodb::basics::parse_dateTime(std::string const& dateTimeIn, tp_sys_clock_ms& date_tp) {
+bool arangodb::basics::parseDateTime(std::string const& dateTimeIn, arangodb::tp_sys_clock_ms& date_tp) {
   using namespace date;
   using namespace std::chrono;
 
@@ -233,8 +775,8 @@ bool arangodb::basics::parse_dateTime(std::string const& dateTimeIn, tp_sys_cloc
   return true;
 }
 
-bool arangodb::basics::regex_isoDuration(std::string const& isoDuration,
-                                         std::smatch& durationParts) {
+bool arangodb::basics::regexIsoDuration(std::string const& isoDuration,
+                                        std::smatch& durationParts) {
   if (isoDuration.length() <= 1) {
     return false;
   }
@@ -245,3 +787,9 @@ bool arangodb::basics::regex_isoDuration(std::string const& isoDuration,
 
   return true;
 }
+
+std::string arangodb::basics::formatDate(std::string const& formatString,
+                                         arangodb::tp_sys_clock_ms const& dateValue) {
+  return ::executeDateFormatRegex(formatString, dateValue);
+}
+
